backend/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
pydantic
Pillow>=10.0.0
